A lot of refactoring and some new features
Removed "installPath" from SlingServerConfiguration to remove the
potential for inconsistencies with SlingBundleConfiguration

Added toString() to SlingServerConfiguration and
SlingServersConfiguration

Since SlingSupport contains a SlingServerConfiguration, removed the
potential for a cyclic relationship by removing the "slingSupport"
property from SlingServerConfiguration

Extracted BUNDLE_CONTROL_BASE_PATH as a constant in
SlingServerConfiguration

Renamed the SlingServerConfiguration.bundleControlUri to
bundlesControlUri (plural) to better reflect that it is not
specific to a particular bundle

Added setTheProperty(..) to SlingServerConfiguration that does
type-coercion and property name translation if needed when setting
the property. Particularly useful when using reflection to configure
the server configuration based on un-typed things like environment
variables.

Removed the "uninstallBundlesPredicate" property from
SlingServersConfiguration until it can find a better home

The SlingServersConfiguration is now much more flexible in how it is configured:

* If project properties have been defined (typically via the
-P command line options) starting with "slingserver." then use
that to override any System properties or values found in the
environment configuration file.

* "slingserver.env.*" has special semantics: see its documentation

* Any other value after the "slingserver." is taken to be the
name of the configuration, which then is followed by the
property to override

* If there isn't a server by that name, it creates a
"default" localhost configuration

* If the name contains "auth" in it then the default port
is set to 4502, otherwise it's assumed to be a publisher
and set to 4503

* For example:

* "slingserver.author.port": 4602
  * Changes the port of the server named "author" to be 4602
* "slingserver.author_2.machineName": test.myco.com
  * Creates a new server configuration (assuming there isn't
    one defined in the environment configuration file) with
    a port of 4502 and machineName of test.myco.com
* "slingserver.twinkle_2.machineName": test.myco.com
  * Creates a new server configuration (assuming there
    isn't one defined in the environment configuration
    file) with a port of 4503 and machineName of
    test.myco.com

Made the SimpleHttpClient much more self-contained for SlingSupport
so it doesn't need to be passed around

Updated SlingSupport to use the SlingPostServlet more, which
simplified a number of the calls (removing the need to
support MKCOL and DELETE, for example)

Better error reporting in SlingSupport

Removed the "slingServers" and "felixId" properties from
SlingBundleConfiguration

The "uploadBundle" task now depends on "deleteBundle"

A number of changes to SlingBundlePlugin to take advantage of
the new APIs

Renamed SlingOsgiBundle to BundleAndServers, and changed most of
its methods to take SlingBundleSupport instead of SlingSupport

BundleAndServers now requires taking in a SlingServersConfiguration.
This allows for it actually act as a bridge between a bundle
configuration and a set of Sling servers.

BundleAndServers.doAcrossServers will only act upon active servers

Reversed the BundleAndServers.doAcrossServers logic to fail if any
call is bad

BundleAndServers.getBundleLocation(..) now understands to difference
between a jcrinstall: URI and an inputstream: URI

Cleaned up the responsibilities for the test fixtures

Added SlingBundlePluginFunctionalSpec that starts an
embedded Jetty server instance

Added SlingSupportFactory (with a SimpleSlingSupportFactory implementation)

Moved all bundle-specific URIs out of SlingServerConfiguration and
into a new class, BundleServerConfiguration

Removed obsolete bundleID lookup in SlingSupport

Added SlingBundleSupport, which is like SlingSupport but adds
bundle-specific functionality

Moved "retryWaitMs" and "maxWaitMs" properties from
SlingServersConfiguration to SlingServerConfiguration

Added the ability to use a "server name" of "env" to specify
properties that should apply across all servers. For example,
setting an environment variable of SLINGSERVERS_ENV_USERNAME=frooble
would default the "username" property of every server configuration
to be "frooble". Setting the same property on a specific server
overrides this default.